ci_template: windows: clone the gst-build specified in the manifest
In all our other builds, we are using the clone_manifest_ref script
to fetch the revision of gst-build that we discover in the manifest.
For the windows job this was missed it seems, but didn't cause
any issues till now cause it only affected the gst-build MRs.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-ci/-/merge_requests/296>