pkgconfig: fix meson warning about waylandlib not being in the config data
meson.build:58: WARNING: The variable(s) 'waylandlibdir' in the input file 'subprojects/gst-plugins-bad/pkgconfig/gstreamer-plugins-bad-uninstalled.pc.in' are not present in the given configuration data.

We don't provide a .pc file for this lib nor install its headers,
so no need for this path to be in the uninstalled .pc file really.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/-/merge_requests/1417>